Add an option for running tests with real modules
For testing purposes, the selinux and semanage modules are mocked. This
allows running the tests without SELinux enabled (e.g. in a container).
On SELinux-enabled systems, it makes sense to rerun the tests with the
actual selinux and semanage modules present on the system. For this
purpose, add a global constant named 'SELINUX_ENABLED'.