Add Dockerfile for udica container
This adds a dockerfile which allows you to build a container for udica
from source.

It also includes a Makefile that'll allow you to build the image with
your preferred command (the default is podman and the default tag is
"latest").

Note that in ordre to use udica as a container, you need to bind-mount
the following directories to it:

* /sys/fs/selinux

* /etc/selinux

* /var/lib/selinux