Refactor tests so they're run from top directory
This makes sure that we don't need to go to the tests/ directory in
order to run the tests. This also has the added effect that you no
longer need to install udica in the system in order to run the tests.

In order to achieve this, we attempt to be clever and determine the path
where the tests are at.

On the other hand, we needed to change the way we overwrite the selinux
python modules. So we used mock.patch instead.