Separate unit from integration tests
We now have a 3 file test structure.

All of the tests are in test_main.py as they've always been.

The unit test call should be done from the test_unit.py module, which is
also called from the Makefile and the CI.

Integration tests (which really call selinux) are called separately and
directly from the test_integration module. This also requires that the
udica package is installed in the calling system.